refactor(request): move organization-request to organization module
