+++ backend/src/modules/organization/constants/organization-filter.config.ts
@@ -16,3 +16,29 @@ export const ORGANIZATION_FILTERS_CONFIG = {
   relations: {},
   rangeColumn: 'createdOn',
 };
+
+export const ORGANIZATION_REQUEST_FILTER_CONFIG = {
+  selectColumns: {
+    id: true,
+    organization: {
+      id: true,
+      organizationGeneral: {
+        name: true,
+      },
+    },
+    email: true,
+    phone: true,
+    name: true,
+    createdOn: true,
+    status: true,
+  },
+  searchableColumns: ['organization.organizationGeneral.name', 'name'],
+  defaultSortBy: 'createdOn',
+  defaultOrderDirection: OrderDirection.ASC,
+  relations: {
+    organization: {
+      organizationGeneral: true,
+    },
+  },
+  rangeColumn: 'createdOn',
+};
