--- backend/src/modules/requests/enums/request-type.enum.ts
@@ -1,4 +1,0 @@
-export enum RequestType {
-  CREATE_ORGANIZATION = 'create_organization',
-  REQUEST_APPLICATION_ACCESS = 'request_application_access',
-}
