+++ backend/src/modules/user/dto/create-user.dto.ts
@@ -12,7 +12,6 @@ import {
 import { REGEX } from 'src/common/constants/patterns.constant';
 import { IsValidPhone } from 'src/common/decorators/validation.decorator';
 import { Role } from '../enums/role.enum';
-import { UserStatus } from '../enums/user-status.enum';
 
 export class CreateUserDto {
   @IsString()
@@ -34,10 +33,6 @@ export class CreateUserDto {
   @IsOptional()
   @IsEnum(Role)
   role?: Role;
-
-  @IsOptional()
-  @IsEnum(UserStatus)
-  status?: UserStatus.PENDING;
 
   @IsNumber()
   organizationId: number;
