+++ backend/src/modules/user/services/cognito.service.ts
@@ -22,11 +22,7 @@ export class CognitoUserService {
     });
   }
 
-  async createUser({
-    name,
-    email,
-    phone,
-  }: Partial<CreateUserDto>): Promise<string> {
+  async createUser({ name, email, phone }: CreateUserDto): Promise<string> {
     const createUserCommand = new AdminCreateUserCommand({
       UserPoolId: CognitoConfig.userPoolId,
       Username: email,
