+++ frontend/src/pages/requests/interfaces/Request.interface.ts
@@ -1,13 +1,13 @@
 import { BaseEntity } from '../../../common/interfaces/base-entity.interface';
 import { Application } from '../../../services/application/interfaces/Application.interface';
 import { IOrganizationFull } from '../../organization/interfaces/Organization.interface';
-import { IUser } from '../../users/interfaces/User.interface';
 import { RequestStatus } from '../enum/RequestStatus.enum';
-import { OngApplication } from './OngApplication.interface';
 
 export interface IOrganizationRequest extends BaseEntity {
   status: RequestStatus;
-  user: IUser;
+  email: string;
+  phone: string;
+  name: string;
   organization: IOrganizationFull;
 }
 
@@ -15,7 +15,6 @@ export interface IApplicationRequest extends BaseEntity {
   status: RequestStatus;
   organizationId: number;
   organization: Partial<IOrganizationFull>;
-
   applicationId: number;
   application: Partial<Application>;
 }
