+++ frontend/src/store/store.ts
@@ -94,7 +94,6 @@ interface ApplicationRequestState {
 interface ApplicationState {
   applications: PaginatedEntity<Application>;
   setApplications: (applications: PaginatedEntity<Application>) => void;
-
   selectedApplication: Application | null;
   setSelectedApplication: (application: Application) => void;
 }
